chore(scanner): fix all cargo clippy warnings
- Auto-fix 35+ warnings with `cargo clippy --fix`
 - Fix empty doc comment line in query/file.rs
 - Refactor unwrap-after-is_some in refresh.rs with if-let / match
 - Use is_ok() instead of `if let Ok(_)` in handlers.rs
 - Use strip_prefix() instead of manual slice in watcher.rs
 - Suppress too_many_arguments for resolve_node_and_fetch_members
   (structural refactoring deferred)

+++ scanner/src/query/buffer.rs
@@ -78,7 +78,7 @@ pub fn parse_buffer(
     if generated_h_line > 0 && suggested_line > generated_h_line { suggested_line = generated_h_line; }
     if last_include_line == 0 {
         for (i, l) in content.lines().enumerate() {
-            if l.contains("#pragma once") { suggested_line = (i + 2) as usize; break; }
+            if l.contains("#pragma once") { suggested_line = i + 2; break; }
         }
     }
